refactor: replace store and redux hooks
