+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './components/App';
 import './styles.css';
 import { Provider } from 'react-redux';
-import { store } from './store/personal-info/personalInfoStore';
+import { store } from './store/store';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
